agents/rag/requirements.txt
# Requirements for RAG Service with ChromaDB
# Install with: pip install -r requirements.txt

chromadb>=0.4.0
flask>=2.0.0
flask-cors>=4.0.0

# Document parsing
pypdf>=4.0.0           # PDF parsing
python-docx>=1.0.0     # Word (.docx) parsing
python-pptx>=0.6.0     # PowerPoint (.pptx) parsing
